refactor: consolidate deploy code in config module
